requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
huggingface_hub==0.19.4
transformers==4.36.0
torch==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
plotly==5.17.0
requests==2.31.0
openai==1.6.0
streamlit-chat==0.1.1